+++ lib/elasticsearch_ex/deserializer.ex
@@ -1,65 +1,75 @@
 defmodule ElasticsearchEx.Deserializer do
+  @moduledoc """
+  Converts a document or source from Elasticsearch data structures into Elixir data structures.
+
+  An example is the convertion of a range `%{"gte" => first, "lte" => last}` into Elixir `Range`
+  or `Date.Range`.
+  """
+
+  ## Module attributes
+
+  @typedoc """
+  Represents the Elasticsearch mappings which is represented by a `Map` with keys as `String`.
+  """
+  @type mappings :: %{required(binary()) => any()}
+
+  @typedoc """
+  Represents a list/stream of documents, a document, a document source or any values.
+  """
+  @type value :: Enumerable.t() | %{required(binary()) => any()}
+
   ## Public functions
 
-  def deserialize_documents(stream, mapping) when is_struct(stream, Stream) do
-    Stream.map(stream, &deserialize_document(&1, mapping))
+  @doc """
+  Deserialize a list of documents, a document or a document source.
+
+  `Stream` is also accepted and returns a `Stream`.
+  """
+  @spec deserialize(value(), mappings()) :: value()
+  def deserialize(value, mappings)
+
+  def deserialize(stream, mapping) when is_struct(stream, Stream) do
+    Stream.map(stream, &deserialize(&1, mapping))
   end
 
-  def deserialize_documents(documents, mapping) when is_list(documents) do
-    Enum.map(documents, &deserialize_document(&1, mapping))
+  def deserialize(values, mapping) when is_list(values) do
+    Enum.map(values, &deserialize(&1, mapping))
   end
 
-  def deserialize_document(%{"_source" => source} = document, mapping) do
-    deserialized_source = deserialize_value(source, mapping)
+  def deserialize(%{"_source" => source} = document, mapping) do
+    deserialized_source = deserialize(source, mapping)
 
     Map.put(document, "_source", deserialized_source)
   end
 
-  @doc """
-  Deserialize a value based on its mapping, if the value doesn't match, the original value is returned.
-  """
-  def deserialize_value(nil, _mapping), do: nil
-
-  def deserialize_value(values, mapping) when is_list(values) do
-    Enum.map(values, &deserialize_value(&1, mapping))
-  end
-
-  def deserialize_value(value, %{"properties" => mapping}) when is_map(value) do
+  def deserialize(value, %{"properties" => mapping}) when is_map(value) do
     Map.new(value, fn {key, value} ->
       key_mapping = Map.fetch!(mapping, key)
-      deserialized_value = deserialize_value(value, key_mapping)
+      deserialized_value = deserialize(value, key_mapping)
 
       {key, deserialized_value}
     end)
   end
 
-  ## Field types
+  def deserialize(blob, %{"type" => "binary"}) when is_binary(blob) do
+    case Base.decode64(blob) do
+      {:ok, value} ->
+        value
 
-  # https://www.elastic.co/guide/en/elasticsearch/reference/current/binary.html
-  def deserialize_value(blob, %{"type" => "binary"}) do
-    if is_binary(blob) do
-      case Base.decode64(blob) do
-        {:ok, value} ->
-          value
-
-        :error ->
-          blob
-      end
+      :error ->
+        blob
     end
   end
 
-  # https://www.elastic.co/guide/en/elasticsearch/reference/current/range.html
-  def deserialize_value(%{"gte" => gte, "lte" => lte}, %{"type" => type})
-      when type in ~w[integer_range long_range] and is_integer(gte) and is_integer(lte) do
+  def deserialize(%{"gte" => gte, "lte" => lte}, %{"type" => type})
+      when type in ~w[integer_range long_range] do
     Range.new(gte, lte)
   end
 
-  # https://www.elastic.co/guide/en/elasticsearch/reference/current/range.html
-  def deserialize_value(%{"gte" => gte, "lte" => lte} = value, %{
+  def deserialize(%{"gte" => gte, "lte" => lte} = value, %{
         "type" => "date_range",
         "format" => "strict_date"
-      })
-      when is_binary(gte) and is_binary(lte) do
+      }) do
     with {:ok, first} <- Date.from_iso8601(gte),
          {:ok, last} <- Date.from_iso8601(lte) do
       Date.range(first, last)
@@ -69,8 +79,7 @@ defmodule ElasticsearchEx.Deserializer do
     end
   end
 
-  # https://www.elastic.co/guide/en/elasticsearch/reference/current/date.html
-  def deserialize_value(value, %{"type" => "date", "format" => "strict_date_time"}) do
+  def deserialize(value, %{"type" => "date", "format" => "strict_date_time"}) do
     case DateTime.from_iso8601(value) do
       {:ok, date_time, 0} ->
         date_time
@@ -80,8 +89,7 @@ defmodule ElasticsearchEx.Deserializer do
     end
   end
 
-  # https://www.elastic.co/guide/en/elasticsearch/reference/current/date.html
-  def deserialize_value(value, %{"type" => "date", "format" => "strict_date"}) do
+  def deserialize(value, %{"type" => "date", "format" => "strict_date"}) do
     case Date.from_iso8601(value) do
       {:ok, date} ->
         date
@@ -91,15 +99,7 @@ defmodule ElasticsearchEx.Deserializer do
     end
   end
 
-  @default_function Application.compile_env(:elasticsearch_ex, :deserializer)
-
-  if @default_function do
-    def deserialize_value(value, mapping) do
-      @default_function.(value, mapping)
-    end
-  else
-    def deserialize_value(value, _mapping) do
-      value
-    end
+  def deserialize(value, _mapping) do
+    value
   end
 end